# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI SaaS Boilerplate",
    "description": "Boilerplate backend para aplicaciones SaaS con FastAPI, AWS Cognito y pagos con Stripe.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Get User Me",
        "description": "Obtiene el perfil completo y configuración del usuario autenticado actual.",
        "operationId": "get_user_me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserMeResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Update User Profile",
        "description": "Actualiza la información del perfil (nombre y apellido).",
        "operationId": "update_user_profile_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/users/me/export": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Export User Data",
        "description": "Exporta todos los datos personales asociados al usuario (GDPR Art. 15 y 20).",
        "operationId": "export_user_data_api_v1_users_me_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDataExport"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/users/me/account": {
      "delete": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Delete User Account",
        "description": "Suprime la cuenta del usuario cumpliendo con el Derecho al Olvido (GDPR Art. 17).\nValida precondiciones (suscripciones activas, organizaciones), anonimiza el perfil,\npreserva las API keys para no romper integraciones de las organizaciones, y encola la baja en el IdP.",
        "operationId": "delete_user_account_api_v1_users_me_account_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDeletionResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/users/settings": {
      "get": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Get User Settings",
        "description": "Obtiene las preferencias de notificación y estado de suscripción del usuario.",
        "operationId": "get_user_settings_api_v1_users_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsRead"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Update User Settings",
        "description": "Actualiza las preferencias de notificaciones.",
        "operationId": "update_user_settings_api_v1_users_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserSettingsRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/users/trial": {
      "post": {
        "tags": [
          "Usuarios"
        ],
        "summary": "Start Trial",
        "description": "Inicia un período de prueba gratuito de 14 días verificando que no haya sido consumido.",
        "operationId": "start_trial_api_v1_users_trial_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountDeletionResponse": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "default": "Tu cuenta y datos personales han sido anonimizados de acuerdo a GDPR."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "anonymized"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Deleted At"
          }
        },
        "type": "object",
        "required": [
          "deleted_at"
        ],
        "title": "AccountDeletionResponse",
        "description": "Confirmación de supresión y anonimización de cuenta bajo GDPR (Derecho al Olvido)."
      },
...
      "ExportAiUsageEvent": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "organization_id",
          "provider",
          "model",
          "status",
          "created_at"
        ],
        "title": "ExportAiUsageEvent"
      },
      "ExportApiKey": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "key_prefix": {
            "type": "string",
            "title": "Key Prefix"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "organization_id",
          "name",
          "key_prefix",
          "scopes",
          "created_at"
        ],
        "title": "ExportApiKey"
      },
      "ExportOrganizationMembership": {
        "properties": {
          "organization_id": {
            "type": "integer",
            "title": "Organization Id"
          },
          "organization_name": {
            "type": "string",
            "title": "Organization Name"
          },
          "organization_slug": {
            "type": "string",
            "title": "Organization Slug"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "joined_at": {
            "type": "string",
            "format": "date-time",
            "title": "Joined At"
          }
        },
        "type": "object",
        "required": [
          "organization_id",
          "organization_name",
          "organization_slug",
          "role",
          "joined_at"
        ],
        "title": "ExportOrganizationMembership"
      },
      "ExportStripeCustomer": {
        "properties": {
          "stripe_customer_id": {
            "type": "string",
            "title": "Stripe Customer Id"
          },
          "subscription_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Status"
          },
          "stripe_subscription_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stripe Subscription Id"
          }
        },
        "type": "object",
        "required": [
          "stripe_customer_id"
        ],
        "title": "ExportStripeCustomer"
      },
      "ExportUserProfile": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cognito_sub": {
            "type": "string",
            "title": "Cognito Sub"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "first_name": {
            "type": "string",
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "title": "Last Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cognito_sub",
          "email",
          "first_name",
          "last_name",
          "created_at"
        ],
        "title": "ExportUserProfile"
      },
      "ExportUserSettings": {
        "properties": {
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "subscription_start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Start Date"
          },
          "subscription_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription End Date"
          },
          "notify_comments": {
            "type": "boolean",
            "title": "Notify Comments",
            "default": false
          },
          "notify_updates": {
            "type": "boolean",
            "title": "Notify Updates",
            "default": false
          },
          "notify_marketing": {
            "type": "boolean",
            "title": "Notify Marketing",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "subscription_status"
        ],
        "title": "ExportUserSettings"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserDataExport": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/ExportUserProfile"
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportUserSettings"
              },
              {
                "type": "null"
              }
            ]
          },
          "memberships": {
            "items": {
              "$ref": "#/components/schemas/ExportOrganizationMembership"
            },
            "type": "array",
            "title": "Memberships"
          },
          "stripe_customer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExportStripeCustomer"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_api_keys": {
            "items": {
              "$ref": "#/components/schemas/ExportApiKey"
            },
            "type": "array",
            "title": "Created Api Keys"
          },
          "ai_usage_events": {
            "items": {
              "$ref": "#/components/schemas/ExportAiUsageEvent"
            },
            "type": "array",
            "title": "Ai Usage Events"
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Exported At"
          }
        },
        "type": "object",
        "required": [
          "user",
          "exported_at"
        ],
        "title": "UserDataExport",
        "description": "Estructura completa de exportación de datos del usuario bajo GDPR (Derecho de Acceso y Portabilidad)."
      },
      "UserMeResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/UserProfileRead"
          },
          "settings": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserSettingsRead"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "profile"
        ],
        "title": "UserMeResponse"
      },
      "UserProfileRead": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name",
            "default": ""
          },
          "last_name": {
            "type": "string",
            "title": "Last Name",
            "default": ""
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "cognito_sub": {
            "type": "string",
            "title": "Cognito Sub"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "cognito_sub",
          "email",
          "created_at"
        ],
        "title": "UserProfileRead"
      },
      "UserProfileUpdate": {
        "properties": {
          "first_name": {
            "type": "string",
            "title": "First Name",
            "default": ""
          },
          "last_name": {
            "type": "string",
            "title": "Last Name",
            "default": ""
          }
        },
        "type": "object",
        "title": "UserProfileUpdate"
      },
      "UserSettingsRead": {
        "properties": {
          "notify_comments": {
            "type": "boolean",
            "title": "Notify Comments",
            "default": false
          },
          "notify_updates": {
            "type": "boolean",
            "title": "Notify Updates",
            "default": false
          },
          "notify_marketing": {
            "type": "boolean",
            "title": "Notify Marketing",
            "default": false
          },
          "subscription_status": {
            "type": "string",
            "title": "Subscription Status"
          },
          "subscription_start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Start Date"
          },
          "subscription_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription End Date"
          },
          "trial_end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trial End Date"
          },
          "is_subscription_active": {
            "type": "boolean",
            "title": "Is Subscription Active"
          },
          "is_trial_active": {
            "type": "boolean",
            "title": "Is Trial Active"
          }
        },
        "type": "object",
        "required": [
          "subscription_status",
          "is_subscription_active",
          "is_trial_active"
        ],
        "title": "UserSettingsRead"
      },
      "UserSettingsUpdate": {
        "properties": {
          "notify_comments": {
            "type": "boolean",
            "title": "Notify Comments",
            "default": false
          },
          "notify_updates": {
            "type": "boolean",
            "title": "Notify Updates",
            "default": false
          },
          "notify_marketing": {
            "type": "boolean",
            "title": "Notify Marketing",
            "default": false
          }
        },
        "type": "object",
        "title": "UserSettingsUpdate"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}